% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komatufte.R
\name{komatufte}
\alias{komatufte}
\title{Rmarkdown Template for koma-scrartcl}
\usage{
komatufte(..., number_sections = TRUE,
  md_extensions = c("-autolink_bare_uris"), csl = NULL,
  colorlinks = TRUE)
}
\arguments{
\item{...}{arguments passed down to \code{\link[bookdown:pdf_document2]{bookdown::pdf_document2()}}}

\item{number_sections}{Whether to number section headers: if \code{TRUE},
figure/table numbers will be of the form \code{X.i}, where \code{X} is the
current first-level section number, and \code{i} is an incremental number
(the i-th figure/table); if \code{FALSE}, figures/tables will be numbered
sequentially in the document from 1, 2, ..., and you cannot cross-reference
section headers in this case.}

\item{md_extensions}{Arguments to be passed to a
specific output format function. For a function \code{foo2()}, its
arguments are passed to \code{foo()}, e.g. \code{...} of
\code{html_document2()} are passed to \code{rmarkdown::html_document()}.}

\item{csl}{bibliography style in the \code{.csl} format}

\item{colorlinks}{should cross-references and links be
colored?}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
This template is a designed for pdf output and acts
as an alternative to the \code{\link[tufte:tufte_handout]{tufte::tufte_handout()}}
template. Whereas that Rmarkdown template is based on
the \href{https://www.ctan.org/pkg/tufte-latex}{tufte-latex}
LaTeX package, owever, \strong{komatufte} is instead
based on a custom template that I have designed using
the KOMA-script scrartcl class. Most of the added functionality
is documented in the \emph{komatufte} vignette
(\code{vignette("komatufte", "komadown")}) and is accessed via the YAML metadata
block of the R Markdown file.
}
\details{
The main features of this template is the use of a
left-aligned column for the main content and a right-aligned
column for marginal content.

This template is more opinionated than \link{scrartcl}
and is not catered to manual tinkering with the style.
}
\examples{

library(rmarkdown)
library(bookdown)
rmarkdown::draft("my-article.Rmd",
                 template = "komatufte",
                 package = "komadown",
                 edit = FALSE)
}
