% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_report.R
\name{lab_report}
\alias{lab_report}
\title{Lab report template}
\usage{
lab_report(..., automark = TRUE, number_sections = TRUE,
  md_extensions = c("-autolink_bare_uris"))
}
\arguments{
\item{...}{Arguments to \code{bookdown::pdf_document2}}

\item{automark}{Uses `scrlayer-scrpage` to include the subsection and section
in the header (upper left).}

\item{number_sections}{Whether to number section headers: if \code{TRUE},
figure/table numbers will be of the form \code{X.i}, where \code{X} is the
current first-level section number, and \code{i} is an incremental number
(the i-th figure/table); if \code{FALSE}, figures/tables will be numbered
sequentially in the document from 1, 2, ..., and you cannot cross-reference
section headers in this case.}

\item{md_extensions}{Arguments to be passed to a
specific output format function. For a function \code{foo2()}, its
arguments are passed to \code{foo()}, e.g. \code{...} of
\code{html_document2()} are passed to \code{rmarkdown::html_document()}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
A lab report template.
}
\examples{

\dontrun{
library(rmarkdown)
library(bookdown)
draft("MyArticle.Rmd", template = "lab_report", package = "rmdtemplates")
}
}
