% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komadown.R
\docType{package}
\name{komadown-package}
\alias{komadown}
\alias{komadown-package}
\title{komadown: R Markdown Templates for the 'KOMA-Script' Classes}
\description{
R Markdown templates based on the 'KOMA-Script' classes for
LaTeX, additionally offering cross-referencing via the 'bookdown' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/komadown}
  \item Report bugs at \url{https://github.com/jolars/komadown/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com} (0000-0002-4029-5945)

}
\keyword{internal}
