% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdtemplates.R
\docType{package}
\name{rmdtemplates-package}
\alias{rmdtemplates}
\alias{rmdtemplates-package}
\title{rmdtemplates: What the Package Does (one line, title case)}
\description{
RMarkdown templates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/rmdtemplates}
  \item Report bugs at \url{https://github.com/jolars/rmdtemplates/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com}

}
\keyword{internal}
